+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 Next
 ----
 
+* Fix a bug where routes with multiple variables (e.g. ``/users/<org>/<user>/posts``) would incorrectly match URLs with extra path segments.
+
 2025.01.13
 ----------